requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
paho-mqtt==1.6.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
